--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/food2Door/DeliveryService.java
@@ -1,5 +1,0 @@
-package com.kodilla.good.patterns.food2Door;
-
-public interface DeliveryService {
-    boolean deliver(Request request);
-}
