--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/food2Door/ShopService.java
@@ -1,5 +1,0 @@
-package com.kodilla.good.patterns.food2Door;
-
-public interface ShopService {
-    boolean checkOrder(Request request);
-}
