--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/food2Door/Supplier.java
@@ -1,5 +1,0 @@
-package com.kodilla.good.patterns.food2Door;
-
-public interface Supplier {
-    boolean process(Request request);
-}
