+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/food2Door/deliveryServices/DeliveryService.java
@@ -1,0 +1,7 @@
+package com.kodilla.good.patterns.food2Door.deliveryServices;
+
+import com.kodilla.good.patterns.food2Door.model.Request;
+
+public interface DeliveryService {
+    boolean deliver(Request request);
+}
