+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/food2Door/model/Order.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.food2Door;
+package com.kodilla.good.patterns.food2Door.model;
 
 public class Order {
     private String productName;
