+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/food2Door/model/User.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.food2Door;
+package com.kodilla.good.patterns.food2Door.model;
 
 public class User {
     private String name;
