+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/food2Door/shopServices/ShopService.java
@@ -1,0 +1,7 @@
+package com.kodilla.good.patterns.food2Door.shopServices;
+
+import com.kodilla.good.patterns.food2Door.model.Request;
+
+public interface ShopService {
+    boolean checkOrder(Request request);
+}
