+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/food2Door/suppliers/Supplier.java
@@ -1,0 +1,7 @@
+package com.kodilla.good.patterns.food2Door.suppliers;
+
+import com.kodilla.good.patterns.food2Door.model.Request;
+
+public interface Supplier {
+    boolean process(Request request);
+}
